+++ src/main/java/com/practice/backend/dao/mapper/IMapper.java
@@ -1,11 +1,13 @@
-package com.practice.backend.mapper;
+package com.practice.backend.dao.mapper;
 
 
 import com.practice.backend.dao.model.IGuapEntity;
+import org.apache.ibatis.annotations.Mapper;
 
 import java.util.List;
 
 /** Общий интерфейс для всех Mappers */
+@Mapper
 public interface IMapper<EntityType extends IGuapEntity> {
 
     List<EntityType> getAll();
