+++ src/main/java/com/practice/backend/dao/service/SectorSettingsMapService.java
@@ -1,6 +1,6 @@
 package com.practice.backend.dao.service;
 
-import com.practice.backend.mapper.ISectorSettingsMapMapper;
+import com.practice.backend.dao.mapper.ISectorSettingsMapMapper;
 import com.practice.backend.dao.model.SectorSettingsMap;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
