+++ src/main/resources/application.yml
@@ -19,7 +19,6 @@ spring:
     username: postgres
     password: teach1postgres
 
-
 mybatis:
-#  type-aliases-package: com.practice.backend.mapper
+  type-aliases-package: com.practice.backend.dao.mapper
   mapper-locations: classpath:**/mapper/*.xml
