+++ src/main/resources/com/practice/backend/dao/mapper/IFeeMapper.xml
@@ -2,7 +2,7 @@
         PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN"
         "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
 
-<mapper namespace="com.practice.backend.mapper.IFeeMapper">
+<mapper namespace="com.practice.backend.dao.mapper.IFeeMapper">
 
     <select id="getAll">
         select * from fee
