+++ src/main/resources/com/practice/backend/dao/mapper/IOperationMapper.xml
@@ -2,7 +2,7 @@
         PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN"
         "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
 
-<mapper namespace="com.practice.backend.mapper.IOperationMapper">
+<mapper namespace="com.practice.backend.dao.mapper.IOperationMapper">
 
     <select id="getAll">
         select * from operation
